+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Release Notes for MailCraft
 
+## 1.0.1 - 2024-07-16
+- Added events for Commerce Order Created and Commerce Order status change
+- Fixed delete function
+
 ## 1.0.0
 - Initial release